# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Industry Evidence Workspace",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents-v1"
        ],
        "summary": "List Documents",
        "operationId": "list_documents_api_v1_documents_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "cutoff",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cutoff"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/theses/{thesis_id}/rerun": {
      "post": {
        "tags": [
          "engine-commands-v1"
        ],
        "summary": "Rerun Assessment",
        "operationId": "rerun_assessment_api_v1_theses__thesis_id__rerun_post",
        "parameters": [
          {
            "name": "thesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thesis Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RerunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/theses/{thesis_id}/propose": {
      "post": {
        "tags": [
          "engine-commands-v1"
        ],
        "summary": "Propose Evidence",
        "description": "Run the propose step for one thesis.\n\nRecall is cutoff-safe (only statements visible before this run started),\nand every proposed link enters the review queue as machine_generated —\nnothing is auto-confirmed.",
        "operationId": "propose_evidence_api_v1_theses__thesis_id__propose_post",
        "parameters": [
          {
            "name": "thesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thesis Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProposeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{document_version_id}/extract": {
      "post": {
        "tags": [
          "engine-commands-v1"
        ],
        "summary": "Extract Statements",
        "description": "Run the extract step over one document version.\n\nAppend-only: statements are added, never replaced.  The engine script\nfeeds only pending versions (spans present, no statements yet); calling\nthis on an already-extracted version will append duplicates.",
        "operationId": "extract_statements_api_v1_documents__document_version_id__extract_post",
        "parameters": [
          {
            "name": "document_version_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Document Version Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/theses/{thesis_id}/causal-steps": {
      "post": {
        "tags": [
          "causal-commands-v1"
        ],
        "summary": "Create Causal Step",
        "operationId": "create_causal_step_api_v1_theses__thesis_id__causal_steps_post",
        "parameters": [
          {
            "name": "thesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thesis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCausalStepRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedCausalStepDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/theses/{thesis_id}/causal-edges": {
      "post": {
        "tags": [
          "causal-commands-v1"
        ],
        "summary": "Create Causal Edge",
        "operationId": "create_causal_edge_api_v1_theses__thesis_id__causal_edges_post",
        "parameters": [
          {
            "name": "thesis_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Thesis Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCausalEdgeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedCausalEdgeDTO"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateCausalEdgeRequest": {
        "properties": {
          "source_step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Step Id"
          },
          "target_step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Step Id"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "creator_type": {
            "type": "string",
            "title": "Creator Type",
            "default": "human"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source_step_id",
          "target_step_id",
          "rationale"
        ],
        "title": "CreateCausalEdgeRequest"
      },
      "CreateCausalStepRequest": {
        "properties": {
          "description": {
            "type": "string",
            "title": "Description"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "description",
          "sequence"
        ],
        "title": "CreateCausalStepRequest"
      },
...
      "CreatedCausalEdgeDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source_step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Source Step Id"
          },
          "target_step_id": {
            "type": "string",
            "format": "uuid",
            "title": "Target Step Id"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "creator_type": {
            "type": "string",
            "title": "Creator Type"
          },
          "review_state": {
            "type": "string",
            "title": "Review State"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "source_step_id",
          "target_step_id",
          "rationale",
          "creator_type",
          "review_state",
          "created_at"
        ],
        "title": "CreatedCausalEdgeDTO"
      },
      "CreatedCausalStepDTO": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "thesis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Thesis Id"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "thesis_id",
          "description",
          "sequence",
          "created_at"
        ],
        "title": "CreatedCausalStepDTO"
      },
...
      "CursorPage": {
        "properties": {
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          },
          "has_more": {
            "type": "boolean",
            "title": "Has More",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CursorPage"
      },
...
      "DocumentListResponse": {
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "v1",
            "title": "Schema Version",
            "default": "v1"
          },
          "basis": {
            "$ref": "#/components/schemas/HistoricalBasisDTO"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/DocumentSummaryDTO"
            },
            "type": "array",
            "title": "Items"
          },
          "page": {
            "$ref": "#/components/schemas/CursorPage"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "basis",
          "items",
          "page"
        ],
        "title": "DocumentListResponse"
      },
      "DocumentSummaryDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content_sha256": {
            "type": "string",
            "title": "Content Sha256"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "available_at": {
            "type": "string",
            "title": "Available At"
          },
          "acquired_at": {
            "type": "string",
            "title": "Acquired At"
          },
          "parser_version": {
            "type": "string",
            "title": "Parser Version"
          },
          "supersedes_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supersedes Id"
          },
          "span_count": {
            "type": "integer",
            "title": "Span Count"
          },
          "statement_count": {
            "type": "integer",
            "title": "Statement Count"
          },
          "parse_state": {
            "type": "string",
            "enum": [
              "parsed",
              "unparsed"
            ],
            "title": "Parse State"
          },
          "extraction_state": {
            "type": "string",
            "enum": [
              "extracted",
              "extracted_empty",
              "failed",
              "not_attempted"
            ],
            "title": "Extraction State"
          },
          "last_extracted_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Extracted At"
          },
          "content_quality": {
            "type": "string",
            "enum": [
              "ok",
              "degenerate",
              "unknown"
            ],
            "title": "Content Quality",
            "default": "unknown"
          },
          "quality_reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Quality Reasons"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "org": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Org"
          },
          "doc_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Doc Kind"
          },
          "entity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "content_sha256",
          "source_url",
          "published_at",
          "available_at",
          "acquired_at",
          "parser_version",
          "supersedes_id",
          "span_count",
          "statement_count",
          "parse_state",
          "extraction_state"
        ],
        "title": "DocumentSummaryDTO"
      },
...
      "ExtractResponse": {
        "properties": {
          "document_version_id": {
            "type": "string",
            "title": "Document Version Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "statement_count": {
            "type": "integer",
            "title": "Statement Count"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "statements": {
            "items": {
              "$ref": "#/components/schemas/ExtractStatementDTO"
            },
            "type": "array",
            "title": "Statements"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_version_id",
          "mode",
          "statement_count",
          "statements"
        ],
        "title": "ExtractResponse",
        "description": "Result of running statement extraction over one document version.\n\nAppend-only: re-running extraction on a version that already has\nstatements will append duplicates; the engine script only feeds\npending versions (spans present, no statements yet).  ``mode`` is\n``mock`` without an LLM key (non-production only).  ``reason`` is the\nhonest explanation when ``statement_count`` is 0 (无片段 / 表格无可提\n事实 / LLM 拒答)."
      },
      "ExtractStatementDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "normalized_text": {
            "type": "string",
            "title": "Normalized Text"
          },
          "observed_period": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Observed Period"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "normalized_text",
          "observed_period"
        ],
        "title": "ExtractStatementDTO",
        "description": "One statement produced by the extraction step."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoricalBasisDTO": {
        "properties": {
          "cutoff": {
            "type": "string",
            "format": "date-time",
            "title": "Cutoff"
          },
          "is_historical": {
            "type": "boolean",
            "title": "Is Historical"
          },
          "ledger_high_watermark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ledger High Watermark"
          },
          "projection_built_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Built At"
          },
          "projection_schema_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Projection Schema Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "cutoff",
          "is_historical"
        ],
        "title": "HistoricalBasisDTO"
      },
...
      "ProposeResponse": {
        "properties": {
          "thesis_id": {
            "type": "string",
            "title": "Thesis Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "link_count": {
            "type": "integer",
            "title": "Link Count"
          },
          "links": {
            "items": {
              "$ref": "#/components/schemas/ProposedLinkDTO"
            },
            "type": "array",
            "title": "Links"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thesis_id",
          "mode",
          "link_count",
          "links"
        ],
        "title": "ProposeResponse",
        "description": "Result of running evidence proposal for one thesis.\n\nEvery proposed link lands in the review queue as\n``machine_generated``; nothing is auto-confirmed.  ``mode`` is\n``mock`` without an LLM key (non-production only)."
      },
      "ProposedLinkDTO": {
        "properties": {
          "link_id": {
            "type": "string",
            "title": "Link Id"
          },
          "source_statement_id": {
            "type": "string",
            "title": "Source Statement Id"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "scope": {
            "additionalProperties": true,
            "type": "object",
            "title": "Scope"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "link_id",
          "source_statement_id",
          "role",
          "reason",
          "scope"
        ],
        "title": "ProposedLinkDTO",
        "description": "One evidence link proposed by the proposer (pending review)."
      },
...
      "RerunAssessmentDTO": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "snapshot_id": {
            "type": "string",
            "title": "Snapshot Id"
          },
          "conclusion": {
            "type": "string",
            "title": "Conclusion"
          },
          "rationale": {
            "type": "string",
            "title": "Rationale"
          },
          "gaps": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Gaps"
          },
          "displayed_as_provisional": {
            "type": "boolean",
            "title": "Displayed As Provisional"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "snapshot_id",
          "conclusion",
          "rationale",
          "gaps",
          "displayed_as_provisional",
          "created_at"
        ],
        "title": "RerunAssessmentDTO"
      },
      "RerunResponse": {
        "properties": {
          "thesis_id": {
            "type": "string",
            "title": "Thesis Id"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "assessment": {
            "$ref": "#/components/schemas/RerunAssessmentDTO"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "thesis_id",
          "mode",
          "assessment"
        ],
        "title": "RerunResponse",
        "description": "Result of re-running the assess step for one thesis.\n\nA rerun freezes a NEW snapshot and appends a NEW provisional assessment;\nprior snapshots/assessments are never touched, and the difference shows\nup in the snapshot-compare view.  ``mode`` is ``mock`` without an LLM key\n(non-production only — production fails closed per provider discipline)."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}